+++ cashweb-secp256k1-sys/build.rs
@@ -26,37 +26,22 @@ extern crate cc;
 use std::env;
 
 fn main() {
-    if cfg!(feature = "external-symbols") {
-        println!("cargo:rustc-link-lib=static=secp256k1");
-        return;
-    }
-
-    // Check whether we can use 64-bit compilation
-    let use_64bit_compilation = if env::var("CARGO_CFG_TARGET_POINTER_WIDTH").unwrap() == "64" {
-        let check = cc::Build::new().file("depend/check_uint128_t.c")
-                                    .cargo_metadata(false)
-                                    .try_compile("check_uint128_t")
-                                    .is_ok();
-        if !check {
-            println!("cargo:warning=Compiling in 32-bit mode on a 64-bit architecture due to lack of uint128_t support.");
-        }
-        check
-    } else {
-        false
-    };
-
     // Actual build
     let mut base_config = cc::Build::new();
-    base_config.include("depend/secp256k1/")
-               .include("depend/secp256k1/include")
-               .include("depend/secp256k1/src")
-               .flag_if_supported("-Wno-unused-function") // some ecmult stuff is defined but not used upstream
-               .define("SECP256K1_BUILD", Some("1"))
-               // TODO these three should be changed to use libgmp, at least until secp PR 290 is merged
-               .define("USE_NUM_NONE", Some("1"))
-               .define("USE_FIELD_INV_BUILTIN", Some("1"))
-               .define("USE_SCALAR_INV_BUILTIN", Some("1"))
-               .define("ENABLE_MODULE_ECDH", Some("1"));
+    base_config
+        .include("depend/secp256k1/")
+        .include("depend/secp256k1/include")
+        .include("depend/secp256k1/src")
+        .flag_if_supported("-Wno-unused-function") // some ecmult stuff is defined but not used upstream
+        .define("SECP256K1_BUILD", Some("1"))
+        .define("ENABLE_MODULE_ECDH", Some("1"))
+        .define("ENABLE_MODULE_SCHNORRSIG", Some("1"))
+        .define("ENABLE_MODULE_EXTRAKEYS", Some("1"))
+        .define("ECMULT_GEN_PREC_BITS", Some("4"))
+        // TODO these three should be changed to use libgmp, at least until secp PR 290 is merged
+        .define("USE_NUM_NONE", Some("1"))
+        .define("USE_FIELD_INV_BUILTIN", Some("1"))
+        .define("USE_SCALAR_INV_BUILTIN", Some("1"));
 
     if cfg!(feature = "lowmemory") {
         base_config.define("ECMULT_WINDOW_SIZE", Some("4")); // A low-enough value to consume neglible memory
@@ -75,22 +60,16 @@ fn main() {
         }
     }
 
-    if use_64bit_compilation {
-        base_config.define("USE_FIELD_5X52", Some("1"))
-                   .define("USE_SCALAR_4X64", Some("1"))
-                   .define("HAVE___INT128", Some("1"));
-    } else {
-        base_config.define("USE_FIELD_10X26", Some("1"))
-                   .define("USE_SCALAR_8X32", Some("1"));
-    }
-
-    if env::var("TARGET").unwrap() == "wasm32-unknown-unknown" {
-        base_config.include("wasm-sysroot");
+    match &env::var("TARGET").unwrap() as &str {
+        "wasm32-unknown-unknown" | "wasm32-wasi" => {
+            base_config.include("wasm-sysroot");
+        }
+        _ => {}
     }
 
     // secp256k1
-    base_config.file("depend/secp256k1/contrib/lax_der_parsing.c")
-               .file("depend/secp256k1/src/secp256k1.c")
-               .compile("libsecp256k1.a");
+    base_config
+        .file("depend/secp256k1/contrib/lax_der_parsing.c")
+        .file("depend/secp256k1/src/secp256k1.c")
+        .compile("libsecp256k1.a");
 }
-